+++ src/components/modal/inputName/eventHandlers.ts
